+++ src/main/java/com/csgomarket/csgomarketapi/payload/response/authentication/UserData.java
@@ -4,13 +4,15 @@ import com.csgomarket.csgomarketapi.model.user.UserSettings;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class UserData {
 
     private String username;
 
-    private long cash;
+    private BigDecimal cash;
 
     private UserSettings userSettings;
 }
